Avrcp: Sending available player CHANGED response during registration
Earlier when BT turns ON then mAvailablePlayerviewchanged flag is set true,
and because of this when we send A2DP START or AVRCP play from DUT
then in first attempt CHANGED response gets ignore and we retured
without updating current play status to remote.
Now we have shifted this logic to available player changed event,
where this flag will be set as false during registration.

Change-Id: I709527ab6667e9b61581471894a526ba59915155
CRs-Fixed: 2238114