MAP: [IOT] Update MAS Email folder name case to small letters
Usecase:
1) Establish MAP MAS EMAIL session.
2) Send message from carkit and wait for sending success
   event report.
3) Observe if Email in carkit is updated as sent or not.
4) Similar case with read status.

Issue:
Email in carkit is not updated as sent.

Fix:
This change updates case of mandatory folder name of MAS Email
account.

Root Cause:
Some remote carkits is case sensitive towards folder
name. So folder name sent in Capital letters is not processed
by this carkit as case of foldername in Folderlisting and Event
Report is not same (telecom/msg/sent for FolderListing and
telecom/msg/SENT for EventReport).

Change-Id: I805c00ebd1e3e4974f8d4a92040a4e124d9f3722
CRs-Fixed: 2272160