AVRCP: Updating proper playback Status to inactive device during SHO
In multi-a2dp scenario after call end, streaming resumes on remote-1(headset)
and we send playback status as PLAYING in CHANGED response. At the same time
if we play song from remote-2(carkit), SHO happens and playback status as PLAYING
goes to remote-2 but we do not update PAUSE to remote-1 although there is only one
active device(remote-2). To handle this discrepancy we are modifying the logic
where if a device is inactive and last changed response to that device is PLAYING
then during SHO we update PAUSE to inactive device.

CRs-Fixed: 2325319
Change-Id: Ia4a64686269842ca3574adf5884756eb4b29482d