Properly convert the remote volume to show the correct Volume level.
When remote sent volume its volume level which should be properly
sync and show correct value on DUT volume bar. But due to use of
math.floor() API it is not returning proper value. So, volume sync
on DUT and remote is mimatched. So to avoid this use the API
math.round() which would return proper values of volume level
to be in sync with remote volume.

CRs-Fixed: 2355158
Change-Id: Idd6fff52dc7a2944d0d08af824505580e299c163