Don't send changed response as playing to remote while already playing
When some remotes reconnect to DUT, it will send Play and we send the
correct play status as playing. After that when FF/Rewind sent from remote
we send immediately changed response as playing. Due to which, remote
is sending Pause and streaming has been paused. So to avoid this don't
send changed response as playing when FF/Rewind has been sent while
already playing.

CRs-Fixed: 2374351
Change-Id: I5c8df3061b30b44920c76f7d45ab8d9ee058e795