TWS_AVRCP: Fix for safe volume not sent to earbud
When multiple devices are connected along with TWS+ earbuds, and
streaming over earbuds with volume above threshold. One of the
earbuds is disconnected while still streaming to earubds or it is
disconnected after switching streaming to other headset. Upon
reconnection of disconnected earbud, this fix ensures stream
volume to be switched to threshold volume.

Change-Id: I5968be68ecf3e78ae2020aa330921536512ab61f