AVRCP: Check if browse service is connected for a player
In device reboot usecases, during browse player initialization,
media browse service is not connected to media player sometimes
because player is not initialized yet. If remote device is
connected, browsing will not work as browse feature bit for the
player will be reset. To fix browsing issue, check if browse
service is connected or retry when setting browse package.

Change-Id: I3592c69118f18cf5f6ed4ef812baaea728e644f2