TWS_AVRCP: Fix for above the threshold volume set after BT off/on
-Fix to set the volume level to threshold if the volume level is
set above threshold during BT toggle. If the non-active earbud
is disconnected last, disconnected active device is updated with
non-active earbud and when a2dp service calls to store the volume
disconnected active device was not checked for TWS+ pair so
the stream volume was stored though it was above threshold.

Change-Id: Ib9d7614c809cec3799f28229d3c1e8bbdb3a55e3